🔀 Merge FirebaseStorageService to FirebaseService and Add helper method and documentation
